+++ tests/agent_server/deploy/scheduler_test_util.py
@@ -21,15 +21,18 @@ import json
 import uuid
 from typing import Collection, Mapping, Set
 
+import utils
 from inmanta.agent import executor
 from inmanta.agent.code_manager import CodeManager
 from inmanta.agent.executor import ExecutorBlueprint, ResourceInstallSpec
 from inmanta.data import ResourceIdStr
 from inmanta.data.model import LEGACY_PIP_DEFAULT, ResourceType
 from inmanta.deploy.state import ResourceDetails
+from inmanta.protocol import Client
 from inmanta.protocol.common import custom_json_encoder
 from inmanta.resources import Id
 from inmanta.types import JsonType
+from utils import retry_limited
 
 
 def make_requires(resources: Mapping[ResourceIdStr, ResourceDetails]) -> Mapping[ResourceIdStr, Set[ResourceIdStr]]:
@@ -88,3 +91,43 @@ class DummyCodeManager(CodeManager):
         self, environment: uuid.UUID, version: int, resource_types: Collection[ResourceType]
     ) -> tuple[Collection[ResourceInstallSpec], executor.FailedResources]:
         return ([ResourceInstallSpec(rt, version, dummyblueprint) for rt in resource_types], {})
+
+
+async def _wait_until_deployment_finishes(client: Client, environment: str, version: int = -1, timeout: int = 10) -> None:
+    """Interface kept for backward compat"""
+
+    async def done():
+        result = await client.resource_list(environment, deploy_summary=True)
+        assert result.code == 200
+        summary = result.result["metadata"]["deploy_summary"]
+        # {'by_state': {'available': 3, 'cancelled': 0, 'deployed': 12, 'deploying': 0, 'failed': 0, 'skipped': 0,
+        #               'skipped_for_undefined': 0, 'unavailable': 0, 'undefined': 0}, 'total': 15}
+        print(summary)
+        available = summary["by_state"]["available"]
+        deploying = summary["by_state"]["deploying"]
+        return available + deploying == 0
+
+    await retry_limited(done, 10)
+
+
+async def wait_full_success(client: Client, environment: str, version: int = -1, timeout: int = 10) -> None:
+    """Interface kept for backward compat"""
+
+    async def done():
+        result = await client.resource_list(environment, deploy_summary=True)
+        assert result.code == 200
+        summary = result.result["metadata"]["deploy_summary"]
+        # {'by_state': {'available': 3, 'cancelled': 0, 'deployed': 12, 'deploying': 0, 'failed': 0, 'skipped': 0,
+        #               'skipped_for_undefined': 0, 'unavailable': 0, 'undefined': 0}, 'total': 15}
+        print(summary)
+        total = summary["total"]
+        success = summary["by_state"]["deployed"]
+        return total == success
+
+    await retry_limited(done, 10)
+
+
+class ClientHelper(utils.ClientHelper):
+
+    async def wait_for_deployed(self, version: int = -1) -> None:
+        await _wait_until_deployment_finishes(self.client, self.environment)
